Fixed failing tests
Some cases that used to fail due to insufficient view/subquery compaction now pass
